refactor(release-lib): extract bundle ownership engine into release_lib.bundles (#240)
Move the deterministic bundle engine -- catalog, dependency resolution,
and the file-ownership rules (template-owned / user-merge / user-skip) --
out of the 703-LOC scripts/apply_bundle.py CLI into release_lib.bundles.
Idempotent file mutation is exactly the logic that must stay tested code,
not model judgment.

- release_lib/bundles.py: canonical engine (everything except main()):
  BUNDLE_DEFINITIONS, resolve_bundles, validate_source/target, the
  copy_tree/copy_files/copy_skip_on_update ownership functions, and the
  merge_pyproject_deps/merge_gitignore mergers.
- scripts/apply_bundle.py is now a thin argparse CLI importing the engine;
  it re-exports BUNDLE_DEFINITIONS so the existing
  `from apply_bundle import ...` unit test stays green, and the subprocess
  CLI/integration tests are unchanged.
- tests/release_lib/test_bundles.py: direct canonical-home coverage
  (composite resolution, catalog/valid-name sync, the .git-worktree-file
  acceptance gotcha).

Verified: full suite -> 425 passed, 2 skipped; ruff clean; CLI --help OK.
